+++ src/styles/Main/UserInfo/InputBox.css
@@ -65,7 +65,7 @@
     justify-content: center;
     align-items: center;
     border-radius: 6px;
-    width: 110px;
+    width: 105px;
     height: 44px;
     background-color: rgb(210, 210, 210);
     color: rgb(114, 114, 114);
