+++ src/styles/Main/UserInfo/SignUp.css
@@ -61,6 +61,7 @@
 }
 
 .sign-up-content-input-box {
+    width: 100%;
     display: flex;
     flex-direction: column;
     gap: 14px;
